option to select user

+++ authit/providers/storage/fileSystem.py
@@ -1,6 +1,7 @@
 import os
 from uuid import uuid4
 from .interface import StorageInterface
+from authit.entities.base import BaseEntity
 import pickle
 
 
@@ -8,9 +9,18 @@ class FileSystemStorage(StorageInterface):
     def __init__(self, config):
         pass
 
-    def create(self, tbl: str, obj: str) -> str:
+    def create(self, tbl: str, obj: BaseEntity) -> str:
         directory = os.path.join("/tmp", tbl)
         os.makedirs(directory, exist_ok=True)
-        with open(f'/tmp/{tbl}/{str(uuid4())}.pkl', 'ab') as dbfile:
+        filePath = os.path.join(directory, f"{obj._id}.pkl")
+        if os.path.exists(filePath):
+            raise ValueError("Data exists")
+        with open(filePath, 'ab') as dbfile:
             pickle.dump(obj, dbfile)
         return f"/tmp/{tbl}"
+    
+    def select(self, tbl: str, key: str):
+        directory = os.path.join("/tmp", tbl)
+        filePath = os.path.join(directory, f"{key}.pkl")
+        with open(filePath, 'rb') as file:
+            return pickle.load(file=file)
